+++ modern-job-board.php
@@ -3,7 +3,7 @@
  * Plugin Name: Modern Job Board
  * Plugin URI: https://github.com/MartinOrton/modern-job-board
  * Description: A lightweight job board plugin for WordPress.
- * Version: 1.3.0
+ * Version: 1.4.0
  * Author: Martin Orton
  * Author URI: https://www.martinorton.com
  * Text Domain: modern-job-board
@@ -14,13 +14,14 @@ if (!defined('ABSPATH')) {
 }
 
 // Define plugin constants.
-define('MJB_VERSION', '1.3.0');
+define('MJB_VERSION', '1.4.0');
 define('MJB_PATH', plugin_dir_path(__FILE__));
 define('MJB_URL', plugin_dir_url(__FILE__));
 
 require_once MJB_PATH . 'includes/class-mjb-resumes.php';
 require_once MJB_PATH . 'includes/class-mjb-activator.php';
 require_once MJB_PATH . 'includes/class-mjb-notices.php';
+require_once MJB_PATH . 'includes/class-mjb-application-guard.php';
 
 register_activation_hook(__FILE__, array('MJB_Activator', 'activate'));
 register_deactivation_hook(__FILE__, array('MJB_Activator', 'deactivate'));